# openapi.json
{
  "components": {
    "responses": {
...
      "InvalidToken": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The token is invalid or expired, or its user is inactive (`\"Invalid token\"`). The UI logs the user out on this response."
      },
...
      "TokenNotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "No `Authorization` header was sent (`\"Token not found\"`)."
      },
...
    },
    "schemas": {
...
      "AccountUserResponse": {
        "properties": {
          "data": {
            "properties": {
              "user": {
                "$ref": "#/components/schemas/AccountUser"
              }
            },
            "required": [
              "user"
            ],
            "type": "object"
          },
          "status": {
            "enum": [
              "success"
            ],
            "type": "string"
          }
        },
        "required": [
          "status",
          "data"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "errors": {
            "properties": {
              "msg": {
                "example": "User not found",
                "type": "string"
              }
            },
            "required": [
              "msg"
            ],
            "type": "object"
          }
        },
        "required": [
          "errors"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "tokenAuth": {
        "description": "The JWT returned by signup/login, sent as the raw header value — no `Bearer ` prefix.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API for Rate It, a social app for rating movies, TV, music, and books. Success responses are `{ status: \"success\", data: { ... } }`; errors are `{ errors: { msg } }`.",
    "title": "Rate It API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/account/complete-profile": {
      "put": {
        "description": "One-time step for accounts created through Google, Facebook, or Apple sign-in, which start without a userName. Optionally sets a password so the user can also log in with email.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "userName": {
                    "type": "string"
                  }
                },
                "required": [
                  "userName"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountUserResponse"
                }
              }
            },
            "description": "The completed account (`isProfileComplete` is true)."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Profile is already complete\"` or `\"This username is already being used\"`."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Finish a social sign-up by choosing a userName",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/account/me": {
      "get": {
        "description": "Works for accounts that have no userName yet (social sign-ups that haven't completed their profile).",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountUserResponse"
                }
              }
            },
            "description": "The caller's account."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Get the logged-in user's account",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/account/picture": {
      "put": {
        "description": "Uploads to Cloudinary and stores the URL on the account, replacing any previous picture.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "picture": {
                    "description": "A JPEG, PNG, or WebP image, at most 5 MB.",
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "picture"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountUserResponse"
                }
              }
            },
            "description": "The account with its new `picture` URL."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No file, a disallowed type (`\"Only JPEG, PNG, and WebP images are allowed\"`), a file over 5 MB (`\"File too large\"`), or an image Cloudinary rejects."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Upload a new profile picture",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/account/resend-verification": {
      "post": {
        "description": "Issues a new link; any earlier link stops working.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Email sent."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"This email is already verified\"`."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Resend the email verification link",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/account/update": {
      "put": {
        "description": "The user always comes from the token; a userName in the body is ignored. Omitted fields are left unchanged.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountUserResponse"
                }
              }
            },
            "description": "The updated account."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Update the logged-in user's name and phone number",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/account/{userName}": {
      "get": {
        "description": "Returns the full `AccountUser` when `userName` is the caller's own, otherwise only the `PublicUser` fields.",
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            },
            "description": "The user's profile."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Get a user's profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/allUsers": {
      "get": {
        "description": "Returns all users in one response, with no paging.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserListResponse"
                }
              }
            },
            "description": "Every user."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "List every user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/friends/follow": {
      "post": {
        "description": "The follower always comes from the token.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "userToFollow": {
                    "description": "userName of the user to follow.",
                    "type": "string"
                  }
                },
                "required": [
                  "userToFollow"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Now following."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"You cannot follow yourself\"` or `\"You already follow this user\"`."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Follow a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/friends/unfollow": {
      "post": {
        "description": "The follower always comes from the token.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "userToUnfollow": {
                    "description": "userName of the user to unfollow.",
                    "type": "string"
                  }
                },
                "required": [
                  "userToUnfollow"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "No longer following."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"You cannot unfollow yourself\"` or `\"You do not currently follow this user\"`."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Unfollow a user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/{userName}/followers": {
      "get": {
        "description": "Users who no longer exist are left out.",
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserListResponse"
                }
              }
            },
            "description": "The users following `userName`."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "List someone's followers",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/{userName}/following": {
      "get": {
        "description": "Users who no longer exist are left out.",
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserListResponse"
                }
              }
            },
            "description": "The users `userName` follows."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "List the users someone follows",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}